refactor(server): declare openapi metadata on the handlers
